requirements.txt
aiohappyeyeballs              2.6.1
aiohttp                       3.12.12
aiosignal                     1.3.2
alabaster                     1.0.0
anyio                         4.9.0
argon2-cffi                   25.1.0
argon2-cffi-bindings          21.2.0
arrow                         1.3.0
astroid                       3.3.10
asttokens                     3.0.0
async-lru                     2.0.5
asyncssh                      2.21.0
atomicwrites                  1.4.1
attrs                         25.3.0
autopep8                      2.0.4
babel                         2.17.0
bcrypt                        4.3.0
beautifulsoup4                4.13.4
binaryornot                   0.4.4
black                         25.1.0
bleach                        6.2.0
certifi                       2025.4.26
cffi                          1.17.1
chardet                       5.2.0
charset-normalizer            3.4.2
click                         8.2.1
cloudpickle                   3.1.1
colorama                      0.4.6
comm                          0.2.2
cookiecutter                  2.6.0
cryptography                  45.0.4
debugpy                       1.8.14
decorator                     5.2.1
defusedxml                    0.7.1
Deprecated                    1.2.18
diff-match-patch              20241021
dill                          0.4.0
docstring-to-markdown         0.17
docutils                      0.21.2
executing                     2.2.0
fastjsonschema                2.21.1
flake8                        7.1.2
fqdn                          1.5.1
frozenlist                    1.7.0
h11                           0.16.0
httpcore                      1.0.9
httpx                         0.28.1
idna                          3.10
imagesize                     1.4.1
importlib_metadata            8.7.0
inflection                    0.5.1
intervaltree                  3.1.0
ipykernel                     6.29.5
ipython                       8.37.0
ipython_pygments_lexers       1.1.1
ipywidgets                    8.1.7
isoduration                   20.11.0
isort                         6.0.1
jaraco.classes                3.4.0
jaraco.context                6.0.1
jaraco.functools              4.1.0
jedi                          0.19.2
jellyfish                     1.2.0
Jinja2                        3.1.6
json5                         0.12.0
jsonpointer                   3.0.0
jsonschema                    4.24.0
jsonschema-specifications     2025.4.1
jupyter                       1.1.1
jupyter_client                8.6.3
jupyter-console               6.6.3
jupyter_core                  5.8.1
jupyter-events                0.12.0
jupyter-lsp                   2.2.5
jupyter_server                2.16.0
jupyter_server_terminals      0.5.3
jupyterlab                    4.4.3
jupyterlab_pygments           0.3.0
jupyterlab_server             2.27.3
jupyterlab_widgets            3.0.15
keyring                       25.6.0
markdown-it-py                3.0.0
MarkupSafe                    3.0.2
matplotlib-inline             0.1.7
mccabe                        0.7.0
mdurl                         0.1.2
mistune                       3.1.3
more-itertools                10.7.0
multidict                     6.4.4
mypy_extensions               1.1.0
nbclient                      0.10.2
nbconvert                     7.16.6
nbformat                      5.10.4
nest-asyncio                  1.6.0
notebook                      7.4.3
notebook_shim                 0.2.4
numpy                         2.2.6
numpydoc                      1.8.0
overrides                     7.7.0
packaging                     25.0
pandocfilters                 1.5.1
parso                         0.8.4
pathspec                      0.12.1
pexpect                       4.9.0
pickleshare                   0.7.5
pip                           25.0.1
platformdirs                  4.3.8
pluggy                        1.6.0
prometheus_client             0.22.1
prompt_toolkit                3.0.51
propcache                     0.3.2
psutil                        7.0.0
ptyprocess                    0.7.0
pure_eval                     0.2.3
pycodestyle                   2.12.1
pycparser                     2.22
pydocstyle                    6.3.0
pyflakes                      3.2.0
PyGithub                      2.6.1
Pygments                      2.19.1
PyJWT                         2.10.1
pylint                        3.3.7
pylint-venv                   3.0.4
pyls-spyder                   0.4.0
PyNaCl                        1.5.0
PyQt5                         5.15.11
PyQt5-Qt5                     5.15.2
PyQt5_sip                     12.17.0
PyQtWebEngine                 5.15.7
PyQtWebEngine-Qt5             5.15.2
python-dateutil               2.9.0.post0
python-json-logger            3.3.0
python-lsp-black              2.0.0
python-lsp-jsonrpc            1.1.2
python-lsp-server             1.12.2
python-slugify                8.0.4
pytoolconfig                  1.3.1
pyuca                         1.2
pywin32                       310
pywin32-ctypes                0.2.3
pywinpty                      2.0.15
PyYAML                        6.0.2
pyzmq                         26.4.0
QDarkStyle                    3.2.3
qstylizer                     0.2.4
QtAwesome                     1.4.0
qtconsole                     5.6.1
QtPy                          2.4.3
referencing                   0.36.2
requests                      2.32.3
rfc3339-validator             0.1.4
rfc3986-validator             0.1.1
rich                          14.0.0
roman-numerals-py             3.1.0
rope                          1.13.0
rpds-py                       0.25.1
rtree                         1.4.0
Send2Trash                    1.8.3
setuptools                    80.9.0
six                           1.17.0
sniffio                       1.3.1
snowballstemmer               3.0.1
sortedcontainers              2.4.0
soupsieve                     2.7
Sphinx                        8.2.3
sphinxcontrib-applehelp       2.0.0
sphinxcontrib-devhelp         2.0.0
sphinxcontrib-htmlhelp        2.1.0
sphinxcontrib-jsmath          1.0.1
sphinxcontrib-qthelp          2.0.0
sphinxcontrib-serializinghtml 2.0.0
spyder                        6.0.7
spyder-kernels                3.0.5
stack-data                    0.6.3
superqt                       0.7.4
tabulate                      0.9.0
terminado                     0.18.1
text-unidecode                1.3
textdistance                  4.6.3
three-merge                   0.1.1
tinycss2                      1.4.0
tomlkit                       0.13.3
tornado                       6.5.1
traitlets                     5.14.3
types-python-dateutil         2.9.0.20250516
typing_extensions             4.14.0
ujson                         5.10.0
uri-template                  1.3.0
urllib3                       2.4.0
watchdog                      6.0.0
wcwidth                       0.2.13
webcolors                     24.11.1
webencodings                  0.5.1
websocket-client              1.8.0
whatthepatch                  1.0.7
widgetsnbextension            4.0.14
wrapt                         1.17.2
yapf                          0.43.0
yarl                          1.20.1
zipp                          3.23.0

numpy~=2.3.1
pandas~=2.3.0
pytz~=2025.2